Fixed the passing of model prediction insiide the priciing result. Updated documentation.

+++ src/main/java/com/wordpress/kkaravitis/pricing/domain/PricingResult.java
@@ -28,5 +28,5 @@ public class PricingResult {
       Double demandMetric;
       Double competitorPrice;
       Double inventoryLevel;
-      Double modelPrediction;
+      BigDecimal modelPrediction;
 }
